refactor: split StrokesGainedPage into stat components
